Update, round (-1+sqrt(-3))/2
这次把全部流程过了一遍，应该没有太大的问题了
具体要不要合并......看其他部分的情况吧

（然而好像代码行尾又乱了？）
