+++ .github/workflows/packer.yml
@@ -6,6 +6,7 @@ on:
     branches: [main]
     paths: 
       - 'config/packer.json'
+      - 'config/fontmap.txt'
       - 'projects/**'
 
 jobs:
