+++ src/Packer/Extensions/ContentExtension.cs
@@ -1,12 +1,10 @@
-﻿using System.Text.RegularExpressions;
-
+﻿using Packer.Models;
 using Serilog;
-using Packer.Models;
 using System;
 using System.IO;
 using System.Linq;
-using System.Collections.Generic;
 using System.Security.Cryptography;
+using System.Text.RegularExpressions;
 
 namespace Packer.Extensions
 {
@@ -40,7 +38,7 @@ namespace Packer.Extensions
 
         /// <summary>
         /// 文本预处理<br></br>
-        /// 目前仅有特殊字符更换，但还是预留了空间
+        /// 目前仅有特殊符号更换，但还是预留了空间
         /// </summary>
         /// <param name="content">待处理的文本</param>
         /// <param name="category">文本种类，用于判断是否转义</param>
@@ -48,7 +46,7 @@ namespace Packer.Extensions
         /// <returns></returns>
         public static string Preprocess(this string content, FileCategory category, Config config)
         {
-            // 特殊符号（元素名称等）替换
+            // 特殊符号替换
             foreach (var mapping in config.CharatcerReplacement)
             {
                 if (content.Contains(mapping.Key))
@@ -92,22 +90,11 @@ namespace Packer.Extensions
         /// <returns></returns>
         public static bool IsSkippedLang(this string location, Config config)
         {
-            foreach(var lang in config.SkippedLanguages)
+            foreach (var lang in config.SkippedLanguages)
             {
                 if (location.Contains(lang, StringComparison.OrdinalIgnoreCase)) return true;
             }
             return false;
-        }
-
-        /// <summary>
-        /// 计算给定流中全体内容的MD5值。
-        /// </summary>
-        /// <param name="stream">被计算的流</param>
-        /// <returns></returns>
-        public static string ComputeMD5(this Stream stream)
-        {
-            stream.Seek(0, SeekOrigin.Begin);
-            return Convert.ToHexString(MD5.Create().ComputeHash(stream));
         }
     }
 }
