+++ src/Packer/Utils.cs
@@ -1,14 +1,14 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
-using System.Text.Json;
-using System.Threading.Tasks;
-using System;
-using System.Security.Cryptography;
 using System.Linq;
 using System.Text.Encodings.Web;
+using System.Text.Json;
+using System.Threading.Tasks;
 
 using Packer.Models;
 using Serilog;
+
 namespace Packer
 {
     static class Utils
@@ -47,7 +47,7 @@ namespace Packer
             var mcmeta = $"./projects/{version}/pack.mcmeta";
             var meta = JsonSerializer.Deserialize<McMeta>(File.ReadAllText(mcmeta));
             var time = DateTime.UtcNow.AddHours(8);
-            meta.Pack.Description += $"\\n打包时间：{time:yyyy-MM-ddTHH:mm:ssZ}";
+            meta.Pack.Description += $"\n打包时间：{time:yyyy-MM-ddTHH:mm:ssZ}";
             var result = JsonSerializer.Serialize(meta, new JsonSerializerOptions()
             {
                 Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
